requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
python-multipart==0.0.6
yfinance==0.2.32
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
